chore(tests): Add config tests exercising all global options (#15085)
* Add failing test for merging timezone configs

* Fix merging of configured timezones

The timezone configuration was not an optional value. This means that in a
configuration that had multiple global sections (ie multi-file configuration),
all of those sections would have to have the same timezone setting. This works
_now_ because the default is the "local" time zone, which is what most users
want, but it makes setting the timezone in a single file impossible.

* chore(tests): Add config tests exercising all global options

To catch things like https://github.com/vectordotdev/vector/issues/15071

Just trying to build up enough confidence for 0.25.1 with this. #15078 is tracking adding more
robust unit tests in this area exercising more combinations and verifying merge behavior.

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>

Signed-off-by: Jesse Szwedko <jesse.szwedko@datadoghq.com>
Co-authored-by: Bruce Guenter <bruce.guenter@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>