fix(ci): Correct github.actions.rest to github.rest.actions (#15153)
Correct github.actions.rest to github.rest.actions

In #15142 we introduced the split between the trusted and untrusted regresison
detector workflow. Unfortunately the trusted side can only be tested once it's
merged into master branch and we had a bug. The github-scripts that download
artifacts referenced `github.actions.rest` when it should have read
`github.rest.actions`.

Please note that because of the nature of workflow_run triggered workflows this
change will not be reflected in the PR's workflow runs.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>