chore(dev): Drop the separate `vdev check style` subcommand (#15988)
We now have a `vdev check style` that runs some file format checks on the source 
code, as well as `vdev check fmt` that runs `cargo fmt --check` _as well as_ 
doing the style check (in addition to `vdev fmt` that does style check fixes and
runs `cargo fmt`). This PR removes the redundancy and references to it.

This was previously separated out in #8862 for the purpose of allowing to run 
the style check on all changes but still avoid running `cargo fmt` when not 
necessary. However, `cargo fmt` is quite light, rarely taking more than a second
or two, so I don't see this as a particularly large obstacle.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>