chore(ci): Run shutdown integration test through vdev (#16166)
Set $DOCKER_HOST to a network host that won't exist. This
causes a timeout, but we're just testing to ensure the source shuts down timely,
so that's actually a good test case. This, however, has the problem that we
can't set $DOCKER_HOST outside of running docker itself, as that will break it,
so there needs to be a separate set of environment just set within the runner.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>