chore(ci): Modify postgres integration tests to work with vdev (#16279)
This just needed to modify the permissions changes to skip volumes named in the
compose file. It would be nice to be able to avoid the `postgres_` prefix in the
test YAML to make the volume name line up with what docker creates from the
compose file, but that's a nicety at this point and not required.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>