fix(observability): Fix tagging registered events in sources (#16416)
Registered events have their tags/labels assigned when the event is registered
rather than when the event is emitted. For most sources, this is during the
source build stage. In order to set these tags properly, the builder needs to be
run in the context of the same span that the component is normally run in. This
span was being set up in the wrong place, and so all registered events that were
set up in the build function weren't tagged properly. This commit fixes that
span setup location.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>