chore(config): Replace `Serialize` with new `ToValue` trait in config schemas (#16460)
* Remove unused `Metadata::map_default_value`

* Add ToValue trait to replace Serialize for Configurable

The `Serialize` trait is only used by the configurable system in order to
translate default values into `serde_json::Value`. Since the `Serialize` trait
is not object safe, this causes complications for making the other traits object
safe in preparation for moving to an inventory-style system for registering
components.

This change introduces an object-safe `ToValue` trait that is used to express a
direct conversion from values into a `serde_json::Value`. Where possible, this
is implemented directly, however the fallback of `serde_json::to_value` is used
for types where this is difficult.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>