fix(ci): Declare some Regression Detector experiments erratic (#16671)
* Declare some Regression Detector experiments erratic

In the days when this project used the 'soaks' we had a flag for declaring some
experiments erratic, that is, when they fired we'd not ding a PR. The Regression
Detector didn't expose an easy way to declare some experiments erratic and, for
the most part, that worked alright. Lately we've been seeing what we call the
'wobble problem' which is vaguely co-incident to
https://github.com/DataDog/lading/pull/466 but not entirely explained by
such. Capacity made the problem more noticable, we think.

Anyhow, this commit sets two experiments `syslog_regex_logs2metric_ddmetrics`
and `file_to_blackhole` as 'erratic'. This has the same meaning as in the old
days but is now exposed in a per-experiment configuration file.

REF SMP-333

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

* empty commit to trigger CI

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>

---------

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>