Optimize path export functions (#681)
* Skip a `.collect` when joining to a `String`

* Optimize OwnedValuePath conversion to String by eliminating temporaries

The string output conversion would create several temporary strings and then
join them into a single output string. Instead, start with the output string and
write everything into that.

* Avoid requiring ownership when converting paths to strings

* Flip implementations of Display and to String for OwnedTargetPath

* Avoid cloning fields in `field_to_string` using `Cow`

* Avoid intermediate String when writing out OwnedSegment

* Back out `itertools` feature change

* Fix formatting of coalesce segments

The formatter for `OwnedValuePath` has its own formatter for an `OwnedSegment`
that is a separate implementation, and they format things slightly differently.
In particular, coalesce segments format with a separator of `|` for
`OwnedValuePath` but with ` | ` (i.e. with spaces) for `OwnedSegment`. This
makes a difference in the tests, as the VRL tests expect the space but the unit
tests expect none.
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>