feat(stdlib): Add proto_encode and parse_proto to encode and parse protocol buffers (#739)
* Add proto_encode and parse_proto to encode and parse protocol buffers

parse_proto accepts a bytes value, a proto descriptor file path and a message type
and returns the VRL value as parsed from the proto.

`parse_proto!(base64_decode("data"), "person.desc", "proto.Person")`
`=> {"id": 42, "name": "John Doe"}`

For `encode_proto`, it does the reverse:

`encode_proto!({"id": 42, "name": "John Doe"}, "person.desc", "proto.Person")`
`=> bytes data`

Note that the path to descriptor file and message type are required to be literals. We do this so
that during compilation we load the message descriptor and reuse it for encoding or parsing.

Most of the code was copied over from the protobuf codec support from vector. However, there's one
notable difference. In `parse_proto`, we don't set fields that are not set. This ensures that
encode_proto and parse_proto remain symmetric (decoding the result of parsing will return the
original result).

* Address comments and move common protobuf logic into src/protobuf

* Fix clippy issues

* Add changelog entry

* Update LICENSE-3rdparty.csv

* Fix examples
Co-authored-by: anubhavg-icpl <anubhavg@infopercept.com>